requirements.txt
asgiref==3.5.0
boto3==1.21.35
botocore==1.24.35
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.6.0
gunicorn==20.1.0
jmespath==1.0.0
mysql-connector-python==8.0.28
Pillow==9.0.1
platformdirs==2.5.1
psycopg2==2.9.3
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.13.1
whitenoise==6.0.0
